+++ push_notifications/admin.py
@@ -1,38 +1,64 @@
-from django.contrib import admin
-from django.contrib.auth import get_user_model
+from django.apps import apps
+from django.contrib import admin, messages
 from django.utils.translation import ugettext_lazy as _
+from .gcm import GCMError
 from .models import APNSDevice, GCMDevice, get_expired_tokens
-from django.db import connection
+from .settings import PUSH_NOTIFICATIONS_SETTINGS as SETTINGS
 
-User = get_user_model()
+
+User = apps.get_model(*SETTINGS["USER_MODEL"].split("."))
 
 
 class DeviceAdmin(admin.ModelAdmin):
 	list_display = ("__str__", "device_id", "user", "active", "date_created")
-	search_fields = ("name", "device_id", "user__%s" % (User.USERNAME_FIELD))
 	list_filter = ("active", )
 	actions = ("send_message", "send_bulk_message", "prune_devices", "enable", "disable")
+	raw_id_fields = ("user", )
 
-	def send_message(self, request, queryset):
+	if hasattr(User, "USERNAME_FIELD"):
+		search_fields = ("name", "device_id", "user__%s" % (User.USERNAME_FIELD))
+	else:
+		search_fields = ("name", "device_id")
+
+	def send_messages(self, request, queryset, bulk=False):
+		"""
+		Provides error handling for DeviceAdmin send_message and send_bulk_message methods.
+		"""
 		ret = []
 		errors = []
 		r = ""
+
 		for device in queryset:
 			try:
-				r = device.send_message("Test single notification")
-			except Exception as e:
+				if bulk:
+					r = queryset.send_message("Test bulk notification")
+				else:
+					r = device.send_message("Test single notification")
+				if r:
+					ret.append(r)
+			except GCMError as e:
 				errors.append(str(e))
-			if r:
-				ret.append(r)
+
+			if bulk:
+				break
+
 		if errors:
-			self.message_user(request, _("Some messages could not be processed: %r" % ("\n".join(errors))))
+			self.message_user(request, _("Some messages could not be processed: %r" % (", ".join(errors))), level=messages.ERROR)
 		if ret:
-			self.message_user(request, _("All messages were sent: %s" % ("\n".join(ret))))
+			if not bulk:
+				ret = ", ".join(ret)
+			if errors:
+				msg = _("Some messages were sent: %s" % (ret))
+			else:
+				msg = _("All messages were sent: %s" % (ret))
+			self.message_user(request, msg)
+
+	def send_message(self, request, queryset):
+		self.send_messages(request, queryset)
 	send_message.short_description = _("Send test message")
 
 	def send_bulk_message(self, request, queryset):
-		r = queryset.send_message("Test bulk notification")
-		self.message_user(request, _("All messages were sent: %s" % (r)))
+		self.send_messages(request, queryset, True)
 	send_bulk_message.short_description = _("Send test message in bulk")
 
 	def enable(self, request, queryset):
@@ -57,18 +83,5 @@ class DeviceAdmin(admin.ModelAdmin):
 			d.save()
 
 
-class GCMDeviceAdmin(DeviceAdmin):
-	"""
-	Inherits from DeviceAdmin to handle displaying gcm device as a hex value
-	"""
-	def device_id_hex(self, obj):
-		if connection.vendor in ("mysql", "sqlite") and obj.device_id:
-			return hex(obj.device_id).rstrip("L")
-		else:
-			return obj.device_id
-	device_id_hex.short_description = "Device ID"
-
-	list_display = ("__str__", "device_id_hex", "user", "active", "date_created")
-
 admin.site.register(APNSDevice, DeviceAdmin)
-admin.site.register(GCMDevice, GCMDeviceAdmin)
+admin.site.register(GCMDevice, DeviceAdmin)