+++ push_notifications/api/rest_framework.py
@@ -1,17 +1,13 @@
 from __future__ import absolute_import
 
-from django.db.backends.base.operations import BaseDatabaseOperations
 from rest_framework import permissions
 from rest_framework.serializers import ModelSerializer, ValidationError
-from rest_framework.validators import UniqueValidator
 from rest_framework.viewsets import ModelViewSet
 from rest_framework.fields import IntegerField
 
 from push_notifications.models import APNSDevice, GCMDevice
 from push_notifications.fields import hex_re
-
-
-BIGINT_MAX_VALUE = BaseDatabaseOperations.integer_field_ranges["BigIntegerField"][1]
+from push_notifications.fields import UNSIGNED_64BIT_INT_MAX_VALUE
 
 # Fields
 
@@ -22,11 +18,12 @@ class HexIntegerField(IntegerField):
 	"""
 
 	def to_internal_value(self, data):
-		# validate that value is a hex number
+		# validate hex string and convert it to the unsigned
+		# integer representation for internal use
 		try:
-			data = int(data, 16)
+			data = int(data, 16) if type(data) != int else data
 		except ValueError:
-			raise ValidationError("ValidationError Device ID is not a valid hex number")
+			raise ValidationError("Device ID is not a valid hex number")
 		return super(HexIntegerField, self).to_internal_value(data)
 
 	def to_representation(self, value):
@@ -36,8 +33,11 @@ class HexIntegerField(IntegerField):
 # Serializers
 class DeviceSerializerMixin(ModelSerializer):
 	class Meta:
-		fields = ("name", "registration_id", "device_id", "active", "date_created")
+		fields = ("id", "name", "registration_id", "device_id", "active", "date_created")
 		read_only_fields = ("date_created", )
+
+		# See https://github.com/tomchristie/django-rest-framework/issues/1101
+		extra_kwargs = {"active": {"default": True}}
 
 
 class APNSDeviceSerializer(ModelSerializer):
@@ -46,9 +46,10 @@ class APNSDeviceSerializer(ModelSerializer):
 		model = APNSDevice
 
 	def validate_registration_id(self, value):
-		# iOS device tokens are 256-bit hexadecimal (64 characters)
+		# iOS device tokens are 256-bit hexadecimal (64 characters). In 2016 Apple is increasing
+		# iOS device tokens to 100 bytes hexadecimal (200 characters).
 
-		if hex_re.match(value) is None or len(value) != 64:
+		if hex_re.match(value) is None or len(value) not in (64, 200):
 			raise ValidationError("Registration ID (device token) is invalid")
 
 		return value
@@ -57,29 +58,49 @@ class APNSDeviceSerializer(ModelSerializer):
 class GCMDeviceSerializer(ModelSerializer):
 	device_id = HexIntegerField(
 		help_text="ANDROID_ID / TelephonyManager.getDeviceId() (e.g: 0x01)",
-		style={'input_type': 'text'},
-		required=False
+		style={"input_type": "text"},
+		required=False,
+		allow_null=True
 	)
 
 	class Meta(DeviceSerializerMixin.Meta):
 		model = GCMDevice
 
-		extra_kwargs = {
-			# Work around an issue with validating the uniqueness of
-			# registration ids of up to 4k
-			'registration_id': {
-				'validators': [
-					UniqueValidator(queryset=GCMDevice.objects.all())
-				]
-			}
-		}
+		extra_kwargs = {"id": {"read_only": False, "required": False}}
 
 	def validate_device_id(self, value):
-		# max value for django.db.models.BigIntegerField is 9223372036854775807
-		# make sure the value is in valid range
-		if value > BIGINT_MAX_VALUE:
-			raise ValidationError("ValidationError Device ID is out of range")
+		# device ids are 64 bit unsigned values
+		if value > UNSIGNED_64BIT_INT_MAX_VALUE:
+			raise ValidationError("Device ID is out of range")
 		return value
+
+	def validate(self, attrs):
+		devices = None
+		primary_key = None
+		request_method = None
+
+		if self.initial_data.get("registration_id", None):
+			if self.instance:
+				request_method = "update"
+				primary_key = self.instance.id
+			else:
+				request_method = "create"
+		else:
+			if self.context["request"].method in ["PUT", "PATCH"]:
+				request_method = "update"
+				primary_key = attrs["id"]
+			elif self.context["request"].method == "POST":
+				request_method = "create"
+
+		if request_method == "update":
+			devices = GCMDevice.objects.filter(registration_id=attrs["registration_id"]) \
+				.exclude(id=primary_key)
+		elif request_method == "create":
+			devices = GCMDevice.objects.filter(registration_id=attrs["registration_id"])
+
+		if devices:
+			raise ValidationError({'registration_id': 'This field must be unique.'})
+		return attrs
 
 
 # Permissions