+++ push_notifications/models.py
@@ -1,10 +1,11 @@
 from __future__ import unicode_literals
 
-from django.conf import settings
 from django.db import models
 from django.utils.encoding import python_2_unicode_compatible
 from django.utils.translation import ugettext_lazy as _
+
 from .fields import HexIntegerField
+from .settings import PUSH_NOTIFICATIONS_SETTINGS as SETTINGS
 
 from consultant.models import ConsultantToken
 
@@ -13,7 +14,9 @@ class Device(models.Model):
 	name = models.CharField(max_length=255, verbose_name=_("Name"), blank=True, null=True)
 	active = models.BooleanField(verbose_name=_("Is active"), default=True,
 		help_text=_("Inactive devices will not be sent notifications"))
+
 	user = models.ForeignKey(ConsultantToken, blank=True, null=True)
+
 	date_created = models.DateTimeField(verbose_name=_("Creation date"), auto_now_add=True, null=True)
 
 	class Meta:
@@ -39,7 +42,7 @@ class GCMDeviceQuerySet(models.query.QuerySet):
 			if message is not None:
 				data["message"] = message
 
-			reg_ids = [rec.registration_id for rec in self if rec.active]
+			reg_ids = list(self.filter(active=True).values_list('registration_id', flat=True))
 			return gcm_send_bulk_message(registration_ids=reg_ids, data=data, **kwargs)
 
 
@@ -74,14 +77,14 @@ class APNSDeviceQuerySet(models.query.QuerySet):
 	def send_message(self, message, **kwargs):
 		if self:
 			from .apns import apns_send_bulk_message
-			reg_ids = [rec.registration_id for rec in self if rec.active]
+			reg_ids = list(self.filter(active=True).values_list('registration_id', flat=True))
 			return apns_send_bulk_message(registration_ids=reg_ids, alert=message, **kwargs)
 
 
 class APNSDevice(Device):
 	device_id = models.UUIDField(verbose_name=_("Device ID"), blank=True, null=True, db_index=True,
 		help_text="UDID / UIDevice.identifierForVendor()")
-	registration_id = models.CharField(verbose_name=_("Registration ID"), max_length=64, unique=True)
+	registration_id = models.CharField(verbose_name=_("Registration ID"), max_length=200, unique=True)
 
 	objects = APNSDeviceManager()